- Tela de medicamento agora corretamente usando autocomplete-new para laboratório e princípios ativos
- Tela de adicionar medicamento importando da Anvisa
- 1 teste para adicionar medicamento da Anvisa
- Correção de um bug que eu causei na tela de criar conta quando eu impedi que o EnumDropdownListFor fosse usado com tipos enums.

+++ CerebelloWebRole.Tests/Tests/AuthenticationControllerTests.cs
@@ -214,7 +214,7 @@ namespace CerebelloWebRole.Tests.Tests
                         MedicalEntityId = me.Id,
                         MedicalSpecialtyId = ms.Id,
                         MedicalSpecialtyName = ms.Name,
-                        MedicalEntityJurisdiction = TypeEstadoBrasileiro.RJ,
+                        MedicalEntityJurisdiction = (int) TypeEstadoBrasileiro.RJ,
                     };
                     Mvc3TestHelper.SetModelStateErrors(controller, vm);
                 }
